refactor(index): split commit contracts
